requirements.txt
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.26.0
tqdm>=4.66.0
scikit-learn>=1.4.0
tensorboard>=2.16.0
matplotlib>=3.8.0
pandas>=2.2.0
face_recognition